+++ lib/kl/primitives/symbols.rb
@@ -2,6 +2,7 @@ module Kl
   module Primitives
     module Symbols
       def intern(str)
+        raise Kl::Error, "#{str} is not a string" unless str.kind_of? String
         # 'true' and 'false' are treated specially and return the
         # corresponding booleans
         if str == 'true'