+++ HISTORY.md
@@ -2,7 +2,8 @@
 
 ## Yet to be Released
 ### Bug Fixes
-- `absvector`, `address->`, `<-address`, and `intern` now throw errors catchable by `trap-error` when type violations occur with their arguments.
+- `absvector`, `address->`, `<-address`, `intern`, `n->string`, `pos`, `string->n`, and `tlstr` now throw errors catchable by `trap-error` when type violations occur with their arguments.
+- `tlstr` now raises an error when applied to an empty string
 
 ## 0.3.1 - January 24, 2013
 - No code changes. Updates Gemspec to refer to Shen 8.0.